role admin e deleção/edição de agendamentos

+++ src/main/java/com/src/barberpoint/service/AgendamentoService.java
@@ -5,7 +5,6 @@ import com.src.barberpoint.repository.AgendamentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,21 +22,16 @@ public class AgendamentoService {
         return agendamentoRepository.findById(id);
     }
 
-    public List<Agendamento> findByBarbeiroId(Long barbeiroId) {
-        return agendamentoRepository.findByBarbeiroId(barbeiroId);
-    }
-
     public Agendamento save(Agendamento agendamento) {
         return agendamentoRepository.save(agendamento);
     }
 
-    public void deleteById(Long id) {
-        agendamentoRepository.deleteById(id);
-    }
-
-    public boolean isBarbeiroAvailable(Long barbeiroId, LocalDateTime start, LocalDateTime end) {
-        List<Agendamento> agendamentos = agendamentoRepository.findByBarbeiroIdAndDataHoraInicioBetween(barbeiroId,
-                start, end);
-        return agendamentos.isEmpty();
+    public boolean deleteById(Long id) {
+        if (agendamentoRepository.existsById(id)) {
+            agendamentoRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
